Inline coerceToForm since it has only one call site and it seems very specific. We also avoid coercing shapes that we don't know (the catch all was coercing types such as ZodArray that are not showing up in the types)
