+++ packages/remix-forms/src/coercions.ts
@@ -1,7 +1,5 @@
 import type { inputFromForm } from 'domain-functions'
 import type { ZodTypeAny } from 'zod'
-import { parseDate } from './prelude'
-import type { ShapeInfo } from './shapeInfo'
 import { shapeInfo } from './shapeInfo'
 
 type ParsedQs = Awaited<ReturnType<typeof inputFromForm>>
@@ -65,17 +63,4 @@ function coerceValue(value: Value, shape?: ZodTypeAny) {
   return value
 }
 
-function coerceToForm(value: unknown, shape: ShapeInfo) {
-  const { typeName } = shape
-  if (typeName === 'ZodBoolean') {
-    return Boolean(value) ?? false
-  }
-
-  if (typeName === 'ZodDate') {
-    return parseDate(value as Date | undefined)
-  }
-
-  return String(value ?? '')
-}
-
-export { coerceValue, coerceToForm }
+export { coerceValue }
